+++ frontend/workflows/image_gen/widget.js
@@ -23,7 +23,7 @@ let cfg;
 
 const inFlight = new Map(); // msgId -> AbortController
 
-const pendingEdits = new Map(); // attId -> only the fields actually edited
+const pendingEdits = new Map(); // attId -> edited fields
 
 export function initWidget(sharedConfig) {
   cfg = sharedConfig;
@@ -46,14 +46,8 @@ function editPrompt(el) {
 function savePrompt(el) {
   const attId = Number(el.dataset.attId);
   const fields = document.querySelectorAll(`.image-gen-edit[data-att-id="${attId}"]`);
-  // Only the fields actually on screen, merged over what was already pending: seeding
-  // both with "" recorded a blank edit whenever the panel had been repainted out from
-  // under the change event, and a blank prompt is not a render anyone asked for.
   const edit = { ...(pendingEdits.get(attId) || {}) };
   for (const t of fields) edit[t.dataset.field] = t.value;
-  // A cleared prompt is rejected here rather than saved and sent: the provider would
-  // answer it with a 400 naming the parameter, and the reroll is where that lands.
-  // Dropping the key repaints the box back to the prompt the image was made with.
   const blanked = typeof edit.prompt === "string" && !edit.prompt.trim();
   if (blanked) delete edit.prompt;
   if (Object.keys(edit).length) pendingEdits.set(attId, edit);
@@ -155,9 +149,6 @@ export function attachmentRenderer(ctx) {
     buttons.reroll || buttons.regen ? `<div class="image-gen-actions">${buttons.reroll}${buttons.regen}</div>` : "";
   const pend = pendingEdits.get(att.id);
   const cm = att.consumption_metadata || {};
-  // Only the fields the edit actually carries, against the stored value normalized to
-  // "": an entry may hold one field alone -- a blanked prompt is dropped from it -- and
-  // an absent key is not an edit to announce.
   const edited = (key) => pend && key in pend && pend[key] !== (cm[key] ?? "");
   const pending = edited("prompt") || edited("negative_prompt") ? pend : undefined;
   const details = attachmentDetailsHtml(att, { esc, escAttr, pending });
