+++ tests/conftest.py
@@ -14,26 +14,7 @@ import backend.database.connection as db_connection
 
 @pytest.fixture(scope="session", autouse=True)
 def _never_the_real_database(tmp_path_factory):
-    """Point ``DB_PATH`` away from the developer's own database, for every test.
-
-    ``DB_PATH`` defaults to ``backend/data/app.db`` -- a real, populated install --
-    and only a fixture redirects it. So a test that reaches the database and forgets
-    to take one does not fail; it quietly opens the real file and writes to it. That
-    is not a flaky test, it is data loss, and it is silent: nothing in the output says
-    which database was touched.
-
-    It happened here. An unfixtured test called ``set_workflow_config`` and replaced a
-    live install's entire image_gen slot -- styles, cloud connections, API keys and
-    imported ComfyUI graphs -- and the suite reported nine passes.
-
-    No assertion catches that afterwards, so the guard is arranged in front of it: the
-    session default is a path with no schema, and a test that reaches the database
-    without isolating it fails on a missing table naming this file. Per-test fixtures
-    monkeypatch straight over it, so isolated tests are unaffected.
-
-    Session-scoped and autouse, so it is in place before the first test and before any
-    fixture that reads ``DB_PATH`` in order to restore it later.
-    """
+    """Make an unisolated database access fail instead of touching a real install."""
     db_connection.DB_PATH = str(tmp_path_factory.mktemp("db_guard") / "unisolated-see-tests-conftest.db")
